requirements.txt
odfpy>=1.3.3
